# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/{publicId}/stream": {
      "get": {
        "description": "Streams the owner-scoped ready object and forwards an optional HTTP byte range without buffering storage content in the API.",
        "operationId": "VideoDeliveryController_stream",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public UUID of the ready video.",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entire ready video stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Accept-Ranges": {
                "schema": {
                  "type": "string",
                  "example": "bytes"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": 1024
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "video/mp4"
                }
              }
            }
          },
          "206": {
            "description": "Requested byte range from a ready video stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Accept-Ranges": {
                "schema": {
                  "type": "string",
                  "example": "bytes"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": 100
                }
              },
              "Content-Range": {
                "schema": {
                  "type": "string",
                  "example": "bytes 0-99/1024"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "video/mp4"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller does not own the video channel (VIDEO_ACCESS_DENIED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video was not found (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for delivery (VIDEO_NOT_READY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "416": {
            "description": "Range is malformed or outside the object (RANGE_NOT_SATISFIABLE)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Stream a ready private video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/download": {
      "get": {
        "description": "Streams the owner-scoped ready object as a safe attachment without exposing storage location or credentials.",
        "operationId": "VideoDeliveryController_download",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "Public UUID of the ready video.",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ready video attachment",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Accept-Ranges": {
                "schema": {
                  "type": "string",
                  "example": "bytes"
                }
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "example": "attachment; filename=\"video.mp4\"; filename*=UTF-8''video.mp4"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": 1024
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "video/mp4"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller does not own the video channel (VIDEO_ACCESS_DENIED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video was not found (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for delivery (VIDEO_NOT_READY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Download a ready private video",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}